refactor hero as component, fixes for mobile view for header and hero
